PUBLIC: Expose cohort start date as an additional pipeline parameter
Currently it is not possible to specify the desired customer cohort for
model validation, as only the cohort_end_date parameter is exposed as a
pipeline parameter and the cohort_start_date is always set to the
calibration_start_date. This change provides more flexibility so that
users can specify a cohort that does not necessarily have to exist at
the beginning of the given dataset.

Change-Id: Id26edbda756687ee2750d86e67509816bbd5379d
GitOrigin-RevId: d746172cde4d920b3fa9319f7bcc6283648e7739